GPX-693: Remove configmap and modify RBAC permissions
Removed the unused configmap.yaml as it duplicated data present in deployment.yaml. Also, adjusted RBAC permissions for rbac authorization uniformly across the board to get, list, and watch.

Modified the application and image versions to match updated requirements. Refracted symbols and strings to reflect the changes in collector values.

This was done to eliminate redundancies and enhance application security. Simplified rbac.yaml to make it easier to manage.