Add dependencies to Chart.yaml
Added 'multena-rbac-collector' to the dependencies section in Chart.yaml. This contains the necessary values and configurations for the helm chart. Also, updated the application version to reflect the changes. This change allows to manage RBAC authorization in one location and thus improves the security and maintainability of our Helm chart.