GPX-693: Merge gp-multena-rbac-collector into gp-multena as a feature
This commit brings the gp-multena-rbac-collector into the gp-multena chart as a feature. It all starts with renaming of the files from `gp-multena-rbac-collector/templates/*.yaml` to `gp-multena/templates/rbac-collector/*.yaml`, and updating the references accordingly within the files. Some major changes include enabling collector, and adding new fields under collector in `values.yaml` file of gp-multena. At the same time, it deletes `.helmignore`, `Chart.yaml`, and `values.yaml` of gp-multena-rbc-collector. By doing so, it removes gp-multena-rbac-collector as a standalone service and integrated it under gp-multena.